LoongArch: Make the code generation of the trap pattern configurable
In some applications (notably the Linux kernel), "break 0" is used as a
trap that a handler may be able to recover.  But in GCC the "trap"
pattern is meant to make the program rightfully die instead.

As [1] describes, sometimes it's vital to distinguish between the two
cases.  The kernel developers prefer "break 1" here, but in the
user-space it's better to trigger a SIGILL instead of SIGTRAP as the
latter is more likely used as a application-defined trap.

To support both cases, make the code generation configurable with a new
option.

[1]:https://lore.kernel.org/20250923061722.24457-1-yangtiezhu@loongson.cn

gcc/

	* config/loongarch/genopts/loongarch.opt.in (-mbreak-code=):
	New.
	* config/loongarch/loongarch.opt: Regenerate.
	* config/loongarch/loongarch.md (trap): Separate to a
	define_insn and a define_expand which takes la_break_code.
	* doc/invoke.texi (-mbreak-code=): Document.
	* config/loongarch/loongarch.opt.urls: Regenerate.

gcc/testsuite

	* gcc.target/loongarch/trap-default.c: New test.
	* gcc.target/loongarch/trap-1.c: New test.