c++/modules: Track all static class variables [PR122421]
The linker error in the PR is caused because when a static is defined
out of the class body, it doesn't yet have a definition and so
read_var_def (which would otherwise have noted it) never gets called.

This instead moves the responsibility for noting class-scope variables
to read_class_def.

	PR c++/122421

gcc/cp/ChangeLog:

	* module.cc (trees_in::read_var_def): Don't handle class-scope
	variables anymore.
	(trees_in::read_class_def): Handle them here instead.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/inst-6_a.C: New test.
	* g++.dg/modules/inst-6_b.C: New test.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>
Reviewed-by: Jason Merrill <jason@redhat.com>