dwarf: handle repeated decl with different btf_decl_tags [PR122248]
The check in gen_btf_tag_dies which asserted that if the target DIE
already had an annotation then it must be the same as the one we are
attempting to add was too strict.  It is valid for multiple declarations
of the same object to appear with different decl_tags, in which case the
tags from each are accumulated in DECL_ATTRIBUTES.  The existing
annotation may not be the same as the one being added, since new tags
will be added to the head of the chain.

The proper behavior is to always replace any existing AT_GNU_annotation
to refer to the chain of annotations we have just constructed, whether
the head of that chain is the same or not.

	PR debug/122248

gcc/

	* dwarf2out.cc (gen_btf_tag_dies): Always replace an existing
	AT_GNU_annotation on the target die.

gcc/testsuite/

	* gcc.dg/debug/dwarf2/dwarf-btf-decl-tag-4.c: New.
	* gcc.dg/debug/dwarf2/dwarf-btf-decl-tag-5.c: New.
	* gcc.dg/debug/dwarf2/dwarf-btf-decl-tag-6.c: New.