Ada: Fix visibility issue for child unit declared as instance on homonym
The reproducer is made up of 9 units containing multiple level of nested
instances, but in the end the problem is that the final child unit has
the same name as the parameter in its instantiation, exposing the wrong
manipulation of the homonym chain done in Analyze_Subprogram_Instantiation.

The fix is to replace this manipulation with a call to Remove_Homonym.

gcc/ada/
	PR ada/48039
	* sem_ch12.adb (Analyze_Subprogram_Instantiation): Call
	Remove_Homonym to remove the enclosing package from visibility.