arc: Fix wrong vector ordering on big-endian architecture
V2HI vectors, explicitly or auto-generated, could be stored in memory wrongly
due to endianness. For example in the following c code stores to the struct
are SLP vectorized, causing them to be stored in the wrong order:

struct S {short a; short b;};
s.a = 520;
s.b = -1;

in the split2 pass the following register set:

(const_vector:V2HI [
        (const_int 520 [0x208])
        (const_int -1 [0xffffffffffffffff])
    ])) "smallTest.c":16:9 484 {*movv2hi_insn}

is converted to:

(const_int -65016 [0xffffffffffff0208])) "smallTest.c":16:9 3 {*movsi_insn}

and is then loaded into the struct. For big-endian this is wrong because
the most significant bytes are written first in memory, storing -1 instead of
520 in s.a .
This patch swaps the 2 values in this step if the target is big-endian.
The added test creates a vector of 2 shorts and verifies the order when
it is passed in a register or in memory.

Regtested for arc and big-endian arc.

gcc/ChangeLog:

	* config/arc/simdext.md(movv2hi_insn): Change order for movv2hi
	for big-endian.

gcc/testsuite/ChangeLog:

	* gcc.target/arc/movv2hi-be.c: New test.

Signed-off-by: Loeka Rogge <loeka@synopsys.com>