vect: Relax gather/scatter scale handling.
Similar to the signed/unsigned patch before this one relaxes the
gather/scatter restrictions on scale factors.  The basic idea is that a
natively unsupported scale factor can still be reached by emitting a
multiplication before the actual gather operation.  As before, we need
to make sure that there is no overflow when multiplying.

gcc/ChangeLog:

	* tree-vect-data-refs.cc (struct gather_scatter_config):
	Add scale.
	(vect_gather_scatter_get_configs): Try various scales.
	(vect_gather_scatter_fn_p): Add scale handling.
	(vect_check_gather_scatter): Add scale parameter.
	* tree-vect-stmts.cc (check_load_store_for_partial_vectors):
	Ditto.
	(vect_truncate_gather_scatter_offset): Ditto.
	(vect_use_grouped_gather): Ditto.
	(get_load_store_type): Ditto.
	(vectorizable_store): Scale offset if necessary.
	(vectorizable_load): Ditto.
	* tree-vectorizer.h (struct vect_load_store_data): Add
	supported_scale.
	(vect_gather_scatter_fn_p): Add argument.