libstdc++: Do not include internal headers in tests
For 42319.cc the PR says that <ios> reproduced the problem, so let's
include that instead. We should also use the no_pch option because
otherwise the test is including everything anyway, and so fails to check
that the char_traits.h header can be included in isolation. There's also
no reason to use an explicit -std=gnu++11 option, we can test it for all
modes instead.

For the thread test there's no reason to use <bits/move.h> instead of
the correct header for std::move.

libstdc++-v3/ChangeLog:

	* testsuite/17_intro/headers/c++2011/42319.cc: Include <ios>
	instead of <bits/char_traits.h>. Add no_pch option. Remove
	explicit -std=gnu++11 option.
	* testsuite/30_threads/thread/swap/1.cc: Include <utility>
	instead of <bits/move.h>.