libstdc++: Fix -Wunused-variable from <regex>
In r16-4709-gc55c1de3a9adb2 I meant to use the result of the
static_cast<char> for the rest of the function following it, but I
accidentally used the original variable __ch. This causes
-Wunused-variable warnings for the __c initialized from the cast.

This fixes the rest of the function to use __c instead of __ch.

libstdc++-v3/ChangeLog:

	* include/bits/regex.tcc (regex_traits::value): Use __c instead
	of __ch.