Documentation for -fident and -Qy/-Qn options [PR122243]
I noticed that the comments for -fident in common.opt were garbled,
and its description is confusing; this is classed as a code generation
option rather than a preprocessor option, and it controls emission of all
".ident" directives in the assembly file, not just those inserted by the
"#ident" preprocessor directive.  Also, the -Qy/-Qn options which have the
same effect as -fident/-fno-ident were documented as System V Options when
in fact they are available on all targets.  Fixed thusly.

gcc/ChangeLog
	PR other/122243
	* common.opt: Clean up comments/documentation for -fident.
	* doc/invoke.texi: Move -Qy/-Qn documentation from System V options
	and combine with -fident/-fno-ident entry.